fixed staged prediction
There was a bug causing staged prediction to always return the same predictions at all stages (actually the prediction at the last stage). The issue was that some of the parameters of the constructed distributions at each stage were actually references to the current parameter values (updated iteratively) so as the paramters changed, the past stage distributional predictions were erroneously updating themselves to be the future stage parameter predictions.